Update Writing process
Step 3: Structured free writing, largely exceeding the one page requirement. 
During this process (1) I tried to "write in a way that comes naturally," (2) worked "from a suitable design" (by making an explicit outline and organizing my thoughts around it), (3)  I also tried to place myself in the background while writing this and not offer opinions. I simply gave facts for the current problems and proposed solutions to the ethics in autonomous driving.